refactor: refactor login form component
