+++ pet-finder/src/Components/Forms/LoginForm/index.tsx
@@ -1,4 +1,4 @@
-import { FC } from 'react'
+import { FC, useCallback } from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 
@@ -13,6 +13,7 @@ import {
   exceptionResponse,
   ExceptionMessage,
 } from '../../../requests/constants'
+import { loginFormFields } from '../../../constants/formFields'
 
 import styles from './LoginForm.module.scss'
 
@@ -29,44 +30,40 @@ export const LoginForm: FC<LoginFormProps> = () => {
     resolver: yupResolver(loginSchema(ERROR_MESSAGES)),
   })
 
-  const onSubmit: SubmitHandler<LoginFormType> = async ({
-    email,
-    password,
-  }) => {
-    console.log(email, password)
-    try {
-      const response = await logIn({ email, password })
-      if (!response.ok) {
-        const serverMessage =
-          exceptionResponse[response.status as keyof ExceptionMessage]
-        setError('root.serverError', { message: serverMessage })
-      } else {
-        const responseBody: IResponseBodyLogIn = await response.json()
-        console.log(responseBody)
+  const onSubmit: SubmitHandler<LoginFormType> = useCallback(
+    async ({ email, password }) => {
+      console.log(email, password)
+      try {
+        const response = await logIn({ email, password })
+        if (!response.ok) {
+          const serverMessage =
+            exceptionResponse[response.status as keyof ExceptionMessage]
+          setError('root.serverError', { message: serverMessage })
+        } else {
+          const responseBody: IResponseBodyLogIn = await response.json()
+          console.log(responseBody)
+        }
+      } catch (err) {
+        if (err instanceof Error) {
+          setError('root.serverError', { message: exceptionResponse[500] })
+        }
       }
-    } catch (err) {
-      if (err instanceof Error) {
-        setError('root.serverError', { message: exceptionResponse[500] })
-      }
-    }
-  }
+    },
+    [setError]
+  )
 
   return (
     <form className={styles.loginForm} onSubmit={handleSubmit(onSubmit)}>
-      <StyledInput
-        inputError={errors.email}
-        inputName="email"
-        label="Email"
-        type="email"
-        {...register('email')}
-      />
-      <StyledInput
-        inputError={errors.password}
-        inputName="password"
-        label="Password"
-        type="password"
-        {...register('password')}
-      />
+      {loginFormFields.map((field) => (
+        <StyledInput
+          key={field.name}
+          inputName={field.name}
+          label={field.label}
+          type={field.type}
+          inputError={errors[field.name]}
+          {...register(field.name)}
+        />
+      ))}
       <StyledButton text="log in" type="submit" disabled={!isValid} />
       {errors.root?.serverError && <p>{errors.root?.serverError.message}</p>}
     </form>
