+++ pet-finder/src/Components/StyledInput/index.tsx
@@ -1,22 +1,8 @@
 import { ReactElement, forwardRef } from 'react'
-import { FieldError } from 'react-hook-form'
+
+import { StyledInputProps } from './styledInput.type'
 
 import style from './StyledInput.module.scss'
-
-export type InputName =
-  | 'name'
-  | 'password'
-  | 'email'
-  | 'confirmPassword'
-  | 'city'
-  | 'country'
-
-export interface StyledInputProps {
-  inputError?: FieldError
-  type: 'text' | 'password' | 'email'
-  inputName: InputName
-  label: string
-}
 
 export const StyledInput = forwardRef<HTMLInputElement, StyledInputProps>(
   (
