+++ pet-finder/src/constants/formFields.ts
@@ -1,16 +1,26 @@
-import { InputName } from '../Components/StyledInput'
+import { InputName } from '../Components/StyledInput/styledInput.type'
 
-export interface IRegisterFormFields {
+export interface IFormFields {
   name: InputName
   label: string
   type: 'text' | 'password' | 'email'
 }
 
-export const registerFormFields: IRegisterFormFields[] = [
+export interface ILoginFormFields {
+  name: 'email' | 'password'
+  label: string
+  type: 'text' | 'password' | 'email'
+}
+export const registerFormFields: IFormFields[] = [
   { name: 'email', label: 'Email', type: 'email' },
   { name: 'password', label: 'Password', type: 'password' },
   { name: 'confirmPassword', label: 'Confirm password', type: 'password' },
   { name: 'name', label: 'Name', type: 'text' },
   { name: 'country', label: 'Country', type: 'text' },
   { name: 'city', label: 'City', type: 'text' },
+]
+
+export const loginFormFields: ILoginFormFields[] = [
+  { name: 'email', label: 'Email', type: 'email' },
+  { name: 'password', label: 'Password', type: 'password' },
 ]
